requirements.txt
numpy>=1.23.5
pandas>=1.5.3
Pillow>=9.4.0
torch>=2.0.1
torchvision>=0.15.1
tqdm>=4.66.1
wandb>=0.16.0
